test login, albums
