+++ css/order/order.css
@@ -3,6 +3,7 @@
         width: 15%;
         height: 100%;
         background-color: #333030;
+        padding-bottom: 70px;
         display: block;
         position: fixed;
         z-index: 0;
