+++ css/page-content/page-content.css
@@ -4,7 +4,8 @@
 @media screen and (min-width: 1000px) {
     .page-content {
         width: 70%;
-        height: 90%;
+        height: 100%;
+        padding-bottom: 70px;
         position: fixed;
         z-index: 1;
         top: 0;
@@ -14,7 +15,8 @@
 @media screen and (max-width: 1000px) {
     .page-content {
         width: 85%;
-        height: 90%;
+        height: 100%;
+        padding-bottom: 70px;
         position: fixed;
         z-index: 1;
         top: 0;
@@ -25,7 +27,8 @@
 @media screen and (max-width: 700px) {
     .page-content {
         width: 100%;
-        height: 90%;
+        height: 100%;
+        padding-bottom: 70px;
         position: fixed;
         z-index: 1;
         top: 0;
