+++ css/player/player.css
@@ -6,7 +6,7 @@
 
 .player {
     width: 100%;
-    height: 10%;
+    height: 70px;
     background-color: #333030;
     position: fixed;
     z-index: 10000;
