+++ index.html
@@ -71,24 +71,6 @@
             <h1 class="white-title white-title-size">Очередь</h1>
             <hr>
             <ul class="track__container-mini" id="order-tracks">
-                <li>
-                    <div class="track__card-mini">
-                        <img src="img/test-img.png" class="track__img-mini">
-                        <div class="track__description">
-                            <p class="track__title">Трек</p>
-                            <p class="track__subtitle">Исполнитель</p>
-                        </div> 
-                    </div>
-                </li>
-                <li>
-                    <div class="track__card-mini">
-                        <img src="img/test-img.png" class="track__img-mini">
-                        <div class="track__description">
-                            <p class="track__title">Трек</p>
-                            <p class="track__subtitle">Исполнитель</p>
-                        </div> 
-                    </div>
-                </li>
             </ul>
         </section>
         <main class="page-content">
@@ -112,11 +94,6 @@
         </main>
         <section class="player">
             <div class="player__track" id="curr-track">
-                <img src="img/test-img.png" class="player__track__img">
-                <div class="player__track__description">
-                    <p class="player__track__description__title">Трек</p>
-                    <p class="player__track__description__subtitle">Исполнитель</p>
-                </div>
             </div>
             <div class="player__controls">
                 <div class="player__buttons">
@@ -151,8 +128,7 @@
                 <span class="player__time" id="curr-volume">100%</span>
             </div>
         </section>
-        <script type="module" src="js/player.js"></script>
-        <script type="module" src="js/firebase.js"></script>
+        <script type="module" src="js/search.js"></script>
     </body>
 
 </html>
